add basic get groups on admin

+++ EJournal/ViewModels/AdminViewModels/GetGroupModel.cs
@@ -10,7 +10,7 @@ namespace EJournal.ViewModels.AdminViewModels
         public int Id { get; set; }
         public string Name { get; set; }
         public int CountOfStudents { get; set; }
-        public int AverageMark { get; set; }
+        public double AverageMark { get; set; }
         public string NameOfCurator { get; set; }
     }
 }
